changed ui a bit
moved config color initialization to ui init to not be loading every frame

changed frame graph to lines to make it more readable

changed globals file in the primitives to the one in the engine (it was the EDITOR one)

+++ MyGameEngine/Primitive.cpp
@@ -1,5 +1,5 @@
 #include "Primitive.h"
-#include "..\MyGameEditor\Globals.h"
+#include "Globals.h"
 
 Cube::Cube(): 
 	a(-1, -1, 1),
@@ -15,18 +15,18 @@ Cube::Cube():
 Sphere::Sphere(){
 
     for (int i = 0; i <= num_subdivisions; ++i) {
-        float lat0 = PI * (-0.5f + (i - 1) / num_subdivisions);
-        float z0 = radius * sin(lat0);
-        float zr0 = radius * cos(lat0);
+        double lat0 = PI * (-0.5f + (i - 1) / num_subdivisions);
+        double z0 = radius * sin(lat0);
+        double zr0 = radius * cos(lat0);
 
-        float lat1 = PI * (-0.5f + (i) / num_subdivisions);
-        float z1 = radius * sin(lat1);
-        float zr1 = radius * cos(lat1);
+        double lat1 = PI * (-0.5f + i / num_subdivisions);
+        double z1 = radius * sin(lat1);
+        double zr1 = radius * cos(lat1);
 
         for (int j = 0; j <= num_subdivisions; ++j) {
-            float lng = 2 * PI * (j - 1) / num_subdivisions;
-            float x = cos(lng);
-            float y = sin(lng);
+            double lng = 2 * PI * (j - 1) / num_subdivisions;
+            double x = cos(lng);
+            double y = sin(lng);
 
             vertices.push_back(x * zr0);
             vertices.push_back(y * zr0);
